feat: implement message handling for TradingViewV2 component
- Introduced a new message handling system in TradingViewV2 to manage TradingView API requests.
- Created dedicated handlers for KLine data requests and layout updates, improving code organization and readability.
- Added a custom hook, useTradingViewMessageHandler, to encapsulate message handling logic and context management.
- Refactored the TradingViewV2 component to utilize the new message handling structure, enhancing maintainability and scalability.
