+++ packages/kit/src/components/TradingView/TradingViewV2/TradingViewV2.tsx
@@ -14,11 +14,11 @@ import WebView from '../../WebView';
 import { getTradingViewTimezone } from '../utils/tradingViewTimezone';
 
 import {
-  fetchTradingViewV2DataWithSlicing,
   useAutoKLineUpdate,
   useAutoTokenDetailUpdate,
   useNavigationHandler,
 } from './hooks';
+import { useTradingViewMessageHandler } from './messageHandlers';
 
 import type { ICustomReceiveHandlerData } from './types';
 import type { IWebViewRef } from '../../WebView/types';
@@ -50,7 +50,6 @@ export function TradingViewV2(props: ITradingViewV2Props & WebViewProps) {
   const systemLocale = useLocaleVariant();
   const theme = useThemeVariant();
   const [devSettings] = useDevSettingsPersistAtom();
-  const { handleNavigation } = useNavigationHandler();
 
   const {
     mode,
@@ -61,6 +60,13 @@ export function TradingViewV2(props: ITradingViewV2Props & WebViewProps) {
     symbol,
     decimal,
   } = props;
+
+  const { handleNavigation } = useNavigationHandler();
+  const { customReceiveHandler } = useTradingViewMessageHandler({
+    tokenAddress,
+    networkId,
+    webRef,
+  });
 
   // Determine the URL to use based on dev settings
   const finalTradingViewUrl = useMemo(() => {
@@ -90,121 +96,6 @@ export function TradingViewV2(props: ITradingViewV2Props & WebViewProps) {
     url.searchParams.set('decimal', decimal?.toString());
     return url.toString();
   }, [finalTradingViewUrl, calendars, systemLocale, theme, symbol, decimal]);
-
-  const customReceiveHandler = useCallback(
-    async ({ data }: ICustomReceiveHandlerData) => {
-      // Debug: Log all incoming messages
-      console.log('🔍 TradingView message received:', {
-        scope: data.scope,
-        method: data.method,
-        origin: data.origin,
-        dataKeys: data.data ? Object.keys(data.data) : 'no data',
-      });
-
-      // {
-      //     "scope": "$private",
-      //     "method": "tradingview_getKLineData",
-      //     "origin": "tradingview.onekey.so",
-      //     "data": {
-      //         "method": "tradingview_getHistoryData",
-      //         "resolution": "1D",
-      //         "from": 1724803200,
-      //         "to": 1750809600,
-      //         "firstDataRequest": true
-      //     }
-      // }
-
-      // Handle TradingView private API requests
-      if (
-        data.scope === '$private' &&
-        data.method === 'tradingview_getKLineData'
-      ) {
-        // Safely extract history data with proper type checking
-        const messageData = data.data;
-        if (
-          messageData &&
-          typeof messageData === 'object' &&
-          'method' in messageData &&
-          'resolution' in messageData &&
-          'from' in messageData &&
-          'to' in messageData
-        ) {
-          // Extract properties safely with explicit checks
-          const safeData = messageData as unknown as Record<string, unknown>;
-          const method = safeData.method as string;
-          const resolution = safeData.resolution as string;
-          const from = safeData.from as number;
-          const to = safeData.to as number;
-          const firstDataRequest = safeData.firstDataRequest as boolean;
-
-          console.log('TradingView request received:', {
-            method,
-            resolution,
-            from,
-            to,
-            firstDataRequest,
-            origin: data.origin,
-          });
-
-          // Use combined function to get sliced data
-          try {
-            const kLineData = await fetchTradingViewV2DataWithSlicing({
-              tokenAddress,
-              networkId,
-              interval: resolution,
-              timeFrom: from,
-              timeTo: to,
-            });
-
-            if (webRef.current && kLineData) {
-              webRef.current.sendMessageViaInjectedScript({
-                type: 'kLineData',
-                payload: {
-                  type: 'history',
-                  kLineData,
-                  requestData: messageData,
-                },
-              });
-            }
-          } catch (error) {
-            console.error('Failed to fetch and send kline data:', error);
-          }
-        }
-      }
-
-      // Handle TradingView layout update messages
-      if (
-        data.scope === '$private' &&
-        data.method === 'tradingview_layoutUpdate'
-      ) {
-        console.log('✅ Layout update method matched!');
-        // Safely extract layout data with proper type checking
-        const messageData = data.data;
-        if (
-          messageData &&
-          typeof messageData === 'object' &&
-          'layout' in messageData
-        ) {
-          // Extract layout property safely
-          const safeData = messageData as unknown as Record<string, unknown>;
-          const layoutString = safeData.layout as string;
-
-          console.log('📡 TradingView layout update received:', data);
-
-          try {
-            const parsedLayoutData = JSON.parse(layoutString);
-            console.log('🎨 Layout data parsed successfully:', {
-              keys: Object.keys(parsedLayoutData),
-              timestamp: Date.now(),
-            });
-          } catch (error) {
-            console.error('❌ Failed to parse layout data:', error);
-          }
-        }
-      }
-    },
-    [tokenAddress, networkId],
-  );
 
   useAutoKLineUpdate({
     tokenAddress,
