+++ packages/kit/src/components/TradingView/TradingViewV2/messageHandlers/types.ts
@@ -1,0 +1,27 @@
+import type { IWebViewRef } from '../../../WebView/types';
+import type { ICustomReceiveHandlerData } from '../types';
+
+export interface IKLineDataRequest {
+  method: string;
+  resolution: string;
+  from: number;
+  to: number;
+  firstDataRequest: boolean;
+}
+
+export interface ILayoutUpdateData {
+  layout: string;
+}
+
+export interface IMessageHandlerContext {
+  tokenAddress?: string;
+  networkId?: string;
+  webRef: React.RefObject<IWebViewRef | null>;
+}
+
+export interface IMessageHandlerParams {
+  data: ICustomReceiveHandlerData['data'];
+  context: IMessageHandlerContext;
+}
+
+export type IMessageHandler = (params: IMessageHandlerParams) => Promise<void>;
